# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "效率仪表盘后端API，提供任务、提交、用户、仓库、项目、组织等数据管理接口",
        "title": "Efficiency Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9990",
    "basePath": "/api",
    "paths": {
...
        "/api/v2/repos": {
            "get": {
                "description": "按条件查询仓库列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repos"
                ],
                "summary": "获取仓库列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期(YYYYMMDD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期(YYYYMMDD)",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ReposListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/repos/branches": {
            "get": {
                "description": "根据仓库地址获取所有分支",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repos"
                ],
                "summary": "获取仓库分支列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "仓库地址",
                        "name": "repoAddr",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.RepoBranchesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/repos/detail": {
            "get": {
                "description": "根据仓库地址获取仓库详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repos"
                ],
                "summary": "获取仓库详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "仓库地址",
                        "name": "repoAddr",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支名",
                        "name": "repoBranch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期(YYYYMMDD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期(YYYYMMDD)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.RepoDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "main.RepoBranchesResponse": {
            "type": "object",
            "properties": {
                "branches": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "main.RepoCommitItem": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "commit_ancient_minutes": {
                    "type": "number"
                },
                "commit_ancient_minutes_manual": {
                    "type": "number"
                },
                "commit_ancient_minutes_reason": {
                    "type": "string"
                },
                "commit_ancient_minutes_reason_manual": {
                    "type": "string"
                },
                "commit_id": {
                    "type": "string"
                },
                "commit_real_ai_minutes": {
                    "type": "number"
                },
                "commit_real_ancient_minutes": {
                    "type": "number"
                },
                "commit_real_minutes": {
                    "type": "number"
                },
                "commit_real_minutes_manual": {
                    "type": "number"
                },
                "commit_real_minutes_reason": {
                    "type": "string"
                },
                "commit_real_minutes_reason_manual": {
                    "type": "string"
                },
                "commit_time": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "diff_lines": {
                    "type": "integer"
                },
                "downstream_tokens": {
                    "type": "integer"
                },
                "efficiency_ratio": {
                    "type": "number"
                },
                "git_user_email": {
                    "type": "string"
                },
                "git_user_name": {
                    "type": "string"
                },
                "repo_addr": {
                    "type": "string"
                },
                "repo_branch": {
                    "type": "string"
                },
                "silica": {
                    "type": "number"
                },
                "task_ids": {
                    "type": "string",
                    "example": "[\"task1\"]"
                },
                "task_ids_silica": {
                    "type": "string",
                    "example": "[\"1.0\"]"
                },
                "updated_at": {
                    "type": "string"
                },
                "upstream_tokens": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "work_dir": {
                    "type": "string"
                }
            }
        },
        "main.RepoDetailResponse": {
            "type": "object",
            "properties": {
                "branches": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "commits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.RepoCommitItem"
                    }
                },
                "efficiency": {
                    "$ref": "#/definitions/main.RepoEfficiency"
                },
                "repo_addr": {
                    "type": "string"
                },
                "repo_branch": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/main.RepoSummary"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.StatTask"
                    }
                }
            }
        },
        "main.RepoEfficiency": {
            "type": "object",
            "properties": {
                "efficiency_ratio": {
                    "type": "number"
                },
                "repo_ancient_minutes": {
                    "type": "number"
                },
                "repo_ancient_minutes_reason": {
                    "type": "string"
                },
                "repo_real_minutes": {
                    "type": "number"
                },
                "repo_real_minutes_reason": {
                    "type": "string"
                }
            }
        },
...
        "main.RepoListItem": {
            "type": "object",
            "properties": {
                "commit_count": {
                    "type": "integer"
                },
                "efficiency_ratio": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "repo_addr": {
                    "type": "string"
                },
                "repo_branch": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "sum_ancient_minutes": {
                    "type": "number"
                },
                "sum_real_minutes": {
                    "type": "number"
                },
                "task_count": {
                    "type": "integer"
                }
            }
        },
        "main.RepoSummary": {
            "type": "object",
            "properties": {
                "commit_count": {
                    "type": "integer"
                },
                "task_count": {
                    "type": "integer"
                }
            }
        },
        "main.ReposListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.RepoListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "main.StatTask": {
            "type": "object",
            "properties": {
                "caller": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "client_ide": {
                    "type": "string"
                },
                "client_os": {
                    "type": "string"
                },
                "client_os_version": {
                    "type": "string"
                },
                "client_version": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "diff_lines": {
                    "type": "integer"
                },
                "downstream_tokens": {
                    "type": "integer"
                },
                "efficiency_ratio": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "repo_addr": {
                    "type": "string"
                },
                "repo_branch": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "task_ancient_minutes": {
                    "type": "number"
                },
                "task_ancient_minutes_manual": {
                    "type": "number"
                },
                "task_ancient_minutes_reason": {
                    "type": "string"
                },
                "task_ancient_minutes_reason_manual": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "task_real_minutes": {
                    "type": "number"
                },
                "task_real_minutes_manual": {
                    "type": "number"
                },
                "task_real_minutes_reason": {
                    "type": "string"
                },
                "task_real_minutes_reason_manual": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upstream_tokens": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "work_dir": {
                    "type": "string"
                },
                "work_dir_id": {
                    "type": "string"
                }
            }
        },
...
    }
}